initial, empty commit for a KeePass 2.x copy
It looks like there's no available repo of all development history of
KeePass 2.x, but the source code for each release is available in zip
files.

This repo will use the files and datetimes from
https://sourceforge.net/projects/keepass/files/KeePass%202.x/
to make it easier to see changes between releases.

The author name and email are for Dominik's sourceforge account,
which is used to publish the app.

The files all match https://github.com/cappert/keepass2, which seems
to serve the same purpose as this repo.